Database updates, testing queries
Updated the sqlite database to have the userID as a primary key so it does not insert duplicates. I also created a test command that uses the fetchone command to select a row based off of the username and then prints the demons in the column. This will be used to get the current list of demons a user has so we can check and see if the demon is already in their compendium